requirements.txt
python-dotenv>=1.0.0
google-generativeai>=0.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.11
flask>=3.0.0
